SOC-2602 | Renaming mixin and changing scroll on reply editor
