+++ front/main/app/components/discussion-inline-reply-editor.js
@@ -1,10 +1,17 @@
-import DiscussionReplyStickyComponentMixin from '../mixins/discussion-reply-sticky-component';
+import DiscussionStickyComponentMixin from '../mixins/discussion-sticky-component';
 import DiscussionInlineEditor from './discussion-inline-editor';
 
 export default DiscussionInlineEditor.extend(
-	DiscussionReplyStickyComponentMixin,
+	DiscussionStickyComponentMixin,
 	{
 		classNames: ['discussion-inline-reply-editor'],
 		containerClassname: '.discussion-inline-editor-floating-container',
+
+		/**
+		 * @returns {void}
+		 */
+		scrollAfterEntityAdded() {
+			Ember.$('html, body').animate({scrollTop: Ember.$('body').height()});
+		},
 	}
 );
