+++ front/main/app/mixins/discussion-sticky-component.js
